Split iterator facade interface in concrete iterator types
This fixes an issue starting from gcc 14.3 where std::sized_sentinel_for
would evaluate to true for the champs iterator.

std::sized_sentinel_for is a concept, and concept expressions are
"analyzed" in an unevaluated context, e.g. like what's contained in
decltype. As such, the following would compile for any iterator_facade,
even if the proper static assert was defined in the body of the
corresponding friend function or method:

```
  using it_t = typename immer::map<int,int>::iterator;
  using diff_t = decltype(std::declval<it_t>() - std::declval<it_t>());
  static_assert(std::is_same_v<diff_t, std::ptrdiff>);
```

The problem with the above is that operator- will not actually compile
for an immer::map iterator in an evaluated context, as it will fail on a
static_assert. But std::sized_sentinel_for cannot see this.

This ended up preventing doing things such as:

1. calling std::ranges::distance on a immer::map iterator pair
2. constructing a std::vector from a immer::map iterator pair
3. calling std::ranges::copy with a immer::map as the range argument

Splitting the iterator_facade into specific types allows us to only
declare those methods / functions that are well defined for this
specific types.